More stdio formatted output <-> argument alignments in *printf().
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk/bogofilter@4227 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ src/main.c
@@ -62,7 +62,7 @@ int main(int argc, char **argv) /*@globals errno,stderr,stdout@*/
     case RC_UNSURE:	exitcode = EX_UNSURE;	break;
     case RC_OK:		exitcode = EX_OK;	break;
     default:
-	fprintf(dbgout, "Unexpected status code - %d\n", status);
+	fprintf(dbgout, "Unexpected status code - %d\n", (int)status);
 	exit(EX_ERROR);
     }
